services/vision-worker/requirements.txt
# Pin loosely; worker tested on Python 3.11.
ultralytics>=8.0.196,<9
opencv-python-headless>=4.8.0,<5
supabase>=2.3.0,<3
python-dotenv>=1.0.0,<2
numpy>=1.23.0,<3
